Add a bunch of modifications based on CR
Mainly:
* Change factories to converters
* Better implementation for ShopifyVariant image uploads
* Better naming for certain methods
* Use prepend instead of class_eval for model decorators
